--- docs/api/reporter.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.reporter`
----------------------------
-
-.. automodule:: certbot.reporter
-   :members:
